Change clash counter to ULL

+++ src/wtsne.hpp
@@ -181,7 +181,7 @@ wtsne_init(const std::vector<uint64_t> &I, const std::vector<uint64_t> &J,
 
 template <typename real_t>
 inline void update_progress(const uint64_t iter, const uint64_t maxIter,
-                            const real_t eta, const real_t Eq, const uint64_t n_clashes) {
+                            const real_t eta, const real_t Eq, const unsigned long long int n_clashes) {
   if (iter % MAX(1, maxIter / 1000) == 0 || iter == maxIter - 1) {
     // Check for keyboard interrupt from python
     if (PyErr_CheckSignals() != 0) {
